Move promise helpers to separate file
