+++ src/lib/readable-stream/generic-reader.ts
@@ -1,6 +1,6 @@
 import assert from '../../stub/assert';
 import { ReadableStream, ReadableStreamCancel, ReadableStreamReader } from '../readable-stream';
-import { newPromise, setPromiseIsHandledToTrue } from '../helpers';
+import { newPromise, setPromiseIsHandledToTrue } from '../helpers/webidl';
 
 export type ReadResult<T> = {
   done: false;
